move instance tracking module file under mapping/semantic
