+++ src/home_robot/home_robot/mapping/semantic/categorical_2d_semantic_map_module.py
@@ -20,7 +20,7 @@ import home_robot.mapping.map_utils as mu
 import home_robot.utils.depth as du
 import home_robot.utils.pose as pu
 import home_robot.utils.rotation as ru
-from home_robot.agent.objectnav_agent.instance_tracking_modules import InstanceMemory
+from home_robot.mapping.semantic.instance_tracking_modules import InstanceMemory
 from home_robot.mapping.semantic.constants import MapConstants as MC
 
 # For debugging input and output maps - shows matplotlib visuals
