+++ src/home_robot/home_robot/mapping/semantic/categorical_2d_semantic_map_state.py
@@ -7,7 +7,7 @@ from typing import Optional
 import numpy as np
 import torch
 
-from home_robot.agent.objectnav_agent.instance_tracking_modules import InstanceMemory
+from home_robot.mapping.semantic.instance_tracking_modules import InstanceMemory
 from home_robot.mapping.map_utils import MapSizeParameters, init_map_and_pose_for_env
 from home_robot.mapping.semantic.constants import MapConstants as MC
